requirements.txt
alabaster==0.7.12
argcomplete==1.9.3
asgiref==3.2.3
Babel==2.8.0
boto3==1.9.234
botocore==1.12.234
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
Django==3.0
django-cors-headers==3.1.0
django-extensions==2.2.6
djangorestframework==3.10.3
djangorestframework-gis==0.15
docutils==0.15.2
durationpy==0.5
firebase-admin==3.0.0
Flask==1.1.1
future==0.16.0
futures==3.1.1
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.19.0
google-resumable-media==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
hjson==3.0.1
httplib2==0.13.1
idna==2.8
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
lxml==4.5.0
MarkupSafe==1.1.1
msgpack==0.6.1
mysqlclient==1.4.6
numpy==1.18.1
packaging==20.0
Pillow==7.0.0
placebo==0.9.0
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
Pygments==2.5.2
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
s3==3.0.0
s3transfer==0.2.1
Shapely==1.7.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.1
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.16.0
wsgi-request-logger==0.4.6
xmltodict==0.12.0
zappa==0.48.2
